+++ eslint-plugin/lib/rules/no-empty-image-src-attribute.js
@@ -69,6 +69,7 @@ module.exports = {
             (attr) => attr.name.name === "src",
           );
           if (srcValue?.value?.value === "") {
+            //to prevent <img src='' alt='Empty image'/>
             context.report({
               node: srcValue,
               messageId: "SpecifySrcAttribute",