+++ test-project/src/vue/avoid-autoplay.vue
@@ -1,4 +1,5 @@
 <template>
-  <video autoplay></video>
-  <video preload="auto"></video>
+  <video autoplay></video> <!-- NonCompliant Avoid autoplay and set preload='none' for video and audio elements. -->
+  <video preload="auto"></video> <!-- NonCompliant Avoid autoplay and set preload='none' for video and audio elements. -->
+  <video preload="none"></video> <!-- Compliant -->
 </template>