+++ test-project/src/vue/no-empty-image-src-attribute.vue
@@ -1,5 +1,5 @@
 <template>
-  <img src="" /> <!-- Non-compliant -->
-  <img /> <!-- Non-compliant -->
+  <img src="" /> <!-- NonCompliant Make sure to specify src attribute when using <img/>. -->
+  <img /> <!-- NonCompliant Make sure to specify src attribute when using <img/>. -->
   <img src="./logo.svg" /> <!-- Compliant -->
 </template>