+++ test-project/src/vue/prefer-lighter-formats-for-image-files.vue
@@ -1,4 +1,4 @@
 <template>
-  <img src="./assets/images/cat.jpg" alt="A cat" /> <!-- Non-compliant -->
+  <img src="./assets/images/cat.jpg" alt="A cat" /> <!-- NonCompliant You should use lighter formats for image files such as webp, avif, svg, jxl -->
   <img src="./assets/images/cat.webp" alt="A cat" /> <!-- Compliant -->
 </template>