+++ test-project/src/vue/prefer-shorthand-css-notations.vue
@@ -1,4 +1,4 @@
 <template>
-  <div style="margin-top: 1em; margin-right: 0; margin-bottom: 2em; margin-left: 0.5em;"></div> <!-- Non-compliant -->
+  <div style="margin-top: 1em; margin-right: 0; margin-bottom: 2em; margin-left: 0.5em;"></div> <!-- NonCompliant Prefer the shorthand CSS notation margin -->
   <div style="margin: 1em 0 2em 0.5em;"></div> <!-- Compliant -->
 </template>